+++ atlasdb-cassandra-multinode-tests/src/test/java/com/palantir/cassandra/multinode/OneNodeDownNodeAvailabilityTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 Palantir Technologies
+ * Copyright 2017 Palantir Technologies
  *
  * Licensed under the BSD-3 License (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,19 +15,18 @@
  */
 package com.palantir.cassandra.multinode;
 
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import com.palantir.atlasdb.keyvalue.api.KeyValueService;
+import com.palantir.atlasdb.keyvalue.api.NodeAvailabilityStatus;
 
-import org.junit.Test;
+public class OneNodeDownNodeAvailabilityTest extends AbstractNodeAvailabilityTest {
 
-import com.google.common.collect.ImmutableMultimap;
-import com.palantir.common.exception.PalantirRuntimeException;
+    @Override
+    protected NodeAvailabilityStatus expectedNodeAvailabilityStatus() {
+        return NodeAvailabilityStatus.QUORUM_AVAILABLE;
+    }
 
-public class OneNodeDownDeleteTest {
-
-    @Test
-    public void deletingThrows() {
-        assertThatThrownBy(() -> OneNodeDownTestSuite.kvs.delete(OneNodeDownTestSuite.TEST_TABLE,
-                ImmutableMultimap.of(OneNodeDownTestSuite.CELL_1_1, OneNodeDownTestSuite.DEFAULT_TIMESTAMP)))
-                .isInstanceOf(PalantirRuntimeException.class);
+    @Override
+    protected KeyValueService getKeyValueService() {
+        return NodesDownTestSetup.kvs;
     }
 }