+++ docs/source/release_notes/release-notes.rst
@@ -42,6 +42,12 @@ develop
     *    - Type
          - Change
 
+
+    *    - |improved|
+         - ``TransactionManager`` now has an API ``getKeyValueServiceStatus`` that can return the health of the underlying KVS. This is designed for applications
+           to implement their availability status taking into account the :ref:`kvs health <kvs-status-check>`
+           (`Pull Request <https://github.com/palantir/atlasdb/pull/1832>`__)
+
     *    - |improved|
          - Improved performance of getRange() on DbKvs. Range requests are now done with a single round trip to the database.
            (`Pull Request <https://github.com/palantir/atlasdb/pull/1805>`__)
@@ -75,6 +81,11 @@ develop
            We have rewritten clients constructed via ``AtlasDbHttpClients`` to account for this new behaviour, but custom clients directly accessing the lock service may be affected.
            This feature is disabled by default, but can be enabled following the TimeLock server configuration :ref:`docs <timelock-server-time-limiting>`.
            (`Pull Request <https://github.com/palantir/atlasdb/pull/1808>`__)
+
+    *    - |deprecated|
+         - ``ConflictDetectionManagers.createDefault(KeyValueService)`` has been deprecated.
+           If you use this method, please replace it with ``ConflictDetectionManagers.create(KeyValueService)``.
+           (`Pull Request 1 <https://github.com/palantir/atlasdb/pull/1822>`__) and (`Pull Request 2 <https://github.com/palantir/atlasdb/pull/1850>`__)
 
     *    - |new|
          - Atlas Console tables now have a join() method.  See ``help("join")`` in Atlas Console for more details.